refactor InRequest pages
