+++ src/App.tsx
@@ -4,7 +4,6 @@ import { AppHeader } from "./components/AppHeader/AppHeader";
 import { Route, Routes } from "react-router-dom";
 import { Home } from "./Home";
 import { Item } from "./Item";
-import { InRequest, INFORMVIEWS } from "./components/InRequest/InRequest";
 import { CheckListTest } from "./components/CheckList/CheckListTest";
 import { ThemeProvider } from "@fluentui/react";
 import { FluentProvider, webLightTheme } from "@fluentui/react-components";
@@ -12,6 +11,7 @@ import { ErrorProvider } from "./providers/ErrorProvider";
 import { ErrorNotification } from "./components/ErrorNotification/ErrorNotification";
 import { UserProvider } from "providers/UserProvider";
 import { Roles } from "components/Roles/Roles";
+import { InRequestNewForm } from "components/InRequest/InRequestNewForm";
 
 function App() {
   return (
@@ -26,10 +26,7 @@ function App() {
                 <Route index element={<Home />} />
                 <Route path="roles" element={<Roles />} />
                 <Route path="item/:itemNum" element={<Item />} />
-                <Route
-                  path="new"
-                  element={<InRequest view={INFORMVIEWS.NEW} />}
-                />
+                <Route path="new" element={<InRequestNewForm />} />
                 <Route path="checklist" element={<CheckListTest />} />
               </Routes>
             </ErrorProvider>
