+++ src/Item.tsx
@@ -2,7 +2,7 @@ import { FunctionComponent } from "react";
 import { useParams } from "react-router-dom";
 import { CheckList } from "./components/CheckList/CheckList";
 import { useChecklistItems } from "./api/CheckListItemApi";
-import { InRequest, INFORMVIEWS } from "./components/InRequest/InRequest";
+import { InRequest } from "./components/InRequest/InRequest";
 
 export const Item: FunctionComponent = (props) => {
   const { itemNum } = useParams();
@@ -11,7 +11,7 @@ export const Item: FunctionComponent = (props) => {
   return (
     <div>
       <h1>Welcome to the Item Page</h1>
-      <InRequest ReqId={Number(itemNum)} view={INFORMVIEWS.COMPACT} />
+      <InRequest ReqId={Number(itemNum)} />
       <CheckList CheckListItems={checklisItems} />
     </div>
   );
